add new tests

+++ src/main/java/com/robertkonrad/blog/dao/PostDAOImpl.java
@@ -32,16 +32,18 @@ public class PostDAOImpl implements PostDAO {
     public void deletePost(int postId) {
         Session session = entityManager.unwrap(Session.class);
         Post post = session.get(Post.class, postId);
-        List<PostTag> postTags = session.createQuery("FROM PostTag pt WHERE pt.post = '" + postId + "'", PostTag.class).getResultList();
-        for (PostTag pt : postTags) {
-            session.delete(pt);
+        if (post != null) {
+            List<PostTag> postTags = session.createQuery("FROM PostTag pt WHERE pt.post = '" + postId + "'", PostTag.class).getResultList();
+            for (PostTag pt : postTags) {
+                session.delete(pt);
+            }
+            if (!post.getImage().equals("")) {
+                String folder = context.getRealPath("/image/");
+                File file = new File(folder + post.getImage());
+                file.delete();
+            }
+            session.delete(post);
         }
-        if (!post.getImage().equals("")) {
-            String folder = context.getRealPath("/image/");
-            File file = new File(folder + post.getImage());
-            file.delete();
-        }
-        session.delete(post);
     }
 
     @Override
